Add filename to chunks; remove MaxTokens display; fix some issues
Added filename because positions aren't necessarily distinct between files, which would cause those chunks' result text to interlace.
Removed MaxTokens because it's just the context window, which is partly freed up as tasks end, so it's weird to show it as if it were a token generation limit.
Fixed a case where the last token would be displayed in the UI twice.
Reduced contextConsumed so it can take advantage of the freed up KV cache, though I ran into an issue with it not fully freeing. Made it able to drop conversations early to prevent *all* the conversations from dying if it exceeds the context size.
Made it clear the file selector upon successful upload so you don't accidentally reupload the same files as easily.
Adjusted system prompt to try to reduce 'natural language' cruft.